feat: Styles configured. Register and Login forms added.
Config/keys file added, which contains the URL to MongoDB cluster.
User file added, which contains the Schema module for user registration
fields. App now has a express a body parser. Routes/users now has
conditionals and checks for registration, but doesn't work properly.
There is a bug that need to be fixed.